fix import and improve clarity
- fixes Qt imports
- fixes circular imports
- improve structure
- improves help docstrings

tested headless on basic example
