+++ ryven-editor/ryven/example_nodes/pkg_test/nodes.py
@@ -1,5 +1,4 @@
-import os
-from ryven.api import on_gui_load
+from ryven.node_env import on_gui_load
 
 from .std import nodes
 from .linalg import nodes
